Remove float32 conversion because it doesn't actually save space after JSON conversion

+++ metaspace/engine/sm/engine/annotation/metrics.py
@@ -282,14 +282,10 @@ def mass_metrics(iso_images_sparse, iso_mzs_sparse, formula_mzs, formula_ints):
     formula_mzs = np.asarray(formula_mzs)
     formula_ints = np.asarray(formula_ints)
     mz_mean, mz_stddev = calc_mz_stddev(iso_images_sparse, iso_mzs_sparse, formula_mzs)
-    # mz_mean is intentionally kept as float64 for the call to calc_mass_errs, as theoretically
-    # some instruments may benefit from this precision. After this calculation, it's only kept for
-    # display purposes, so it's fine to reduce it to float32.
     mz_err_abs, mz_err_rel = calc_mass_errs(mz_mean, formula_mzs, formula_ints)
 
-    # Convert to float32 for smaller storage
-    mz_mean = np.float32(mz_mean)
-    mz_stddev = np.float32(mz_stddev)
-    mz_err_abs = np.float32(mz_err_abs)
-    mz_err_rel = np.float32(mz_err_rel)
+    mz_mean = mz_mean
+    mz_stddev = mz_stddev
+    mz_err_abs = mz_err_abs
+    mz_err_rel = mz_err_rel
     return mz_mean, mz_stddev, mz_err_abs, mz_err_rel
